+++ src/main/java/teamguu/backend/domain/team/service/TeamService.java
@@ -12,6 +12,7 @@ import teamguu.backend.domain.team.dto.SimpleTeamInfoResponseDto;
 import teamguu.backend.domain.team.dto.TeamInfoResponseDto;
 import teamguu.backend.domain.team.entity.Team;
 import teamguu.backend.domain.team.repository.TeamRepository;
+import teamguu.backend.exception.situation.AlreadyBasicImageException;
 import teamguu.backend.exception.situation.TeamNameAlreadyExistsException;
 import teamguu.backend.exception.situation.TeamNotFoundException;
 
@@ -22,7 +23,6 @@ import static java.util.stream.Collectors.toList;
 
 @Service
 @RequiredArgsConstructor
-@Transactional
 public class TeamService {
 
     private final TeamRepository teamRepository;
@@ -35,48 +35,51 @@ public class TeamService {
         teamRepository.save(createTeamRequestDto.toEntity(captain));
     }
 
-    @Transactional(readOnly = true)
     public TeamInfoResponseDto getTeamInfo(Long teamId) {
-        return TeamInfoResponseDto.from(findTeam(teamId));
+        return TeamInfoResponseDto.from(teamRepository.findById(teamId).orElseThrow(TeamNotFoundException::new));
     }
 
-    @Transactional(readOnly = true)
-    public List<SimpleTeamInfoResponseDto> getSimpleTeamInfoList(Member captain) {
+    public List<SimpleTeamInfoResponseDto> getSimpleTeamInfos(Member captain) {
         return teamRepository.findTeamsByCaptain(captain)
                 .stream()
                 .map(SimpleTeamInfoResponseDto::from)
                 .collect(toList());
     }
 
+    @Transactional
     public void editTeamInfo(EditTeamInfoRequestDto editTeamInfoRequestDto, Long teamId) {
-        Team findTeam = findTeam(teamId);
-        validateDuplicateByName(editTeamInfoRequestDto, findTeam);
-        findTeam.editTeam(editTeamInfoRequestDto.getName(), editTeamInfoRequestDto.getHistory(),
+        Team foundTeam = getTeam(teamId);
+        validateDuplicateByName(editTeamInfoRequestDto, foundTeam);
+        foundTeam.editTeam(editTeamInfoRequestDto.getName(), editTeamInfoRequestDto.getHistory(),
                 editTeamInfoRequestDto.getIntro(), editTeamInfoRequestDto.getPlayerInfo());
     }
 
     public void deleteTeam(Long teamId) {
-        Team foundTeam = findTeam(teamId);
+        Team foundTeam = getTeam(teamId);
         deleteLogoImageIfExits(foundTeam);
         teamRepository.delete(foundTeam);
     }
 
+    @Transactional
     public String changeLogoImageToNew(MultipartFile logoImage, Long teamId){
-        Team foundTeam = findTeam(teamId);
-        String uploadedLogoImageUrl = amazonS3Service.uploadFile(logoImage);
+        Team foundTeam = getTeam(teamId);
         deleteLogoImageIfExits(foundTeam);
-        return foundTeam.changeLogoImageUrl(uploadedLogoImageUrl);
+        return foundTeam.changeLogoImageUrl(amazonS3Service.uploadFile(logoImage));
     }
 
+    @Transactional
     public void changeLogoImageToBasic(Long teamId) {
-        Team foundTeam = findTeam(teamId);
+        Team foundTeam = getTeam(teamId);
         String deleteLogoImageUrl = foundTeam.getLogoImageUrl();
+        if (deleteLogoImageUrl.equals("basic_team.png")) {
+            throw new AlreadyBasicImageException();
+        }
+
         foundTeam.changeLogoImageUrl("basic_team.png");
         amazonS3Service.deleteFile(deleteLogoImageUrl);
     }
 
-    @Transactional(readOnly = true)
-    public Team findTeam(Long teamId) {
+    public Team getTeam(Long teamId) {
         return teamRepository.findById(teamId).orElseThrow(TeamNotFoundException::new);
     }
